remove WCMP1 support
