+++ README.md
@@ -8,7 +8,6 @@ pywcmp provides validation and quality assessment capabilities for the [WMO
 WIS](https://community.wmo.int/activity-areas/wis/wis-overview) Core Metadata
 Profile (WCMP).
 
-- validation against [WCMP 1.3](http://wis.wmo.int/2013/metadata/version_1-3-0/WMO_Core_Metadata_Profile_v1.3_Part_1.pdf), specifically [Part 2](http://wis.wmo.int/2013/metadata/version_1-3-0/WMO_Core_Metadata_Profile_v1.3_Part_2.pdf), Section 2, implementing an executable test suite against the ATS
 - validation against [WCMP 2 (draft)](https://github.com/wmo-im/wcmp2), specifically [Annex A: Conformance Class Abstract Test Suite](https://wmo-im.github.io/wcmp2/standard/wcmp2-DRAFT.html#_conformance_class_abstract_test_suite_normative), implementing an executable test suite against the ATS
 - quality assessement against the [WCMP Key Performance Indicators](https://community.wmo.int/activity-areas/wis/wis-metadata-kpis)
 
@@ -48,11 +47,8 @@ pywcmp bundle sync
 
 # abstract test suite
 
-# validate WCMP 1.3 metadata against abstract test suite (file on disk)
-pywcmp ets validate /path/to/file.xml
-
-# validate WCMP 1.3 metadata against abstract test suite (URL)
-pywcmp ets validate https://example.org/path/to/file.xml
+# validate WCMP 2 metadata against abstract test suite (file on disk)
+pywcmp ets validate /path/to/file.json
 
 # validate WCMP 2 metadata against abstract test suite (URL)
 pywcmp ets validate https://example.org/path/to/file.json
@@ -61,24 +57,21 @@ pywcmp ets validate https://example.org/path/to/file.json
 pywcmp ets validate https://example.org/path/to/file.json --no-fail-on-schema-validation
 
 # adjust debugging messages (CRITICAL, ERROR, WARNING, INFO, DEBUG) to stdout
-pywcmp ets validate https://example.org/path/to/file.xml --verbosity DEBUG
+pywcmp ets validate https://example.org/path/to/file.json --verbosity DEBUG
 
 # write results to logfile
 pywcmp ets validate https://example.org/path/to/file.json --verbosity DEBUG --logfile /tmp/foo.txt
 
 # key performance indicators
 
-# all key performance indicators at once # note: running KPIs automatically runs the ets
-pywcmp kpi validate https://example.org/path/to/file.xml --verbosity DEBUG
+# all key performance indicators at once
+pywcmp kpi validate https://example.org/path/to/file.json --verbosity DEBUG
 
 # all key performance indicators at once, in summary
-pywcmp kpi validate https://example.org/path/to/file.xml --verbosity DEBUG --summary
-
-# all key performance indicators at once, with scoring rubric grouping
-pywcmp kpi validate https://example.org/path/to/file.xml --verbosity DEBUG --group
+pywcmp kpi validate https://example.org/path/to/file.json --verbosity DEBUG --summary
 
 # selected key performance indicator
-pywcmp kpi validate --kpi 4 /path/to/file.xml -v INFO
+pywcmp kpi validate --kpi title /path/to/file.json -v INFO
 
 # WIS2 topic hierarchies
 
@@ -90,24 +83,25 @@ pywcmp topics validate wis2.can --fuzzy
 
 # list children of a given WIS2 topic hierarchy level
 pywcmp topics list wis2.a
-
 ```
 
 ## Using the API
 ```pycon
 >>> # test a file on disk
->>> from lxml import etree
+>>> import json
 >>> from pywcmp.ets import ets
->>> exml = etree.parse('/path/to/file.xml')
+>>> 
+>>> with open('/path/to/file.json')) as fh:
+...     data = json.load(fh)
 >>> # test ETS
->>> ts = ets.WMOCoreMetadataProfileTestSuite13(exml)
+>>> ts = ets.WMOCoreMetadataProfileTestSuite2(datal)
 >>> ts.run_tests()  # raises ValueError error stack on exception
 >>> # test a URL
 >>> from urllib2 import urlopen
 >>> from StringIO import StringIO
 >>> content = StringIO(urlopen('https://....').read())
->>> exml = etree.parse(content)
->>> ts = ets.WMOCoreMetadataProfileTestSuite13(exml)
+>>> data = json.loads(content)
+>>> ts = ets.WMOCoreMetadataProfileTestSuite2(data)
 >>> ts.run_tests()  # raises ValueError error stack on exception
 >>> # handle pywcmp.errors.TestSuiteError
 >>> # pywcmp.errors.TestSuiteError.errors is a list of errors
@@ -118,11 +112,9 @@ pywcmp topics list wis2.a
 >>> ...
 >>> # test KPI
 >>> from pywcmp.kpi import WMOCoreMetadataProfileKeyPerformanceIndicators
->>> kpis = WMOCoreMetadataProfileKeyPerformanceIndicators(exml)
+>>> kpis = WMOCoreMetadataProfileKeyPerformanceIndicators(data)
 >>> results = kpis.evaluate()
 >>> results['summary']
->>> # scoring rubric
->>> grouped = group_kpi_results(results)
 ```
 
 ## Development
