+++ pywcmp/bundle.py
@@ -31,7 +31,6 @@ import shutil
 import zipfile
 
 import click
-from lxml import etree
 
 from pywcmp.wcmp2.topics import build_topics, WIS2_TOPIC_HIERARCHY_LOOKUP
 from pywcmp.util import (get_cli_common_options, get_userdir, urlopen_,
@@ -41,7 +40,6 @@ LOGGER = logging.getLogger(__name__)
 
 USERDIR = get_userdir()
 
-WCMP1_FILES = get_userdir() / 'wcmp-1.3'
 WCMP2_FILES = get_userdir() / 'wcmp-2'
 WIS2_TOPIC_HIERARCHY_DIR = get_userdir() / 'wis2-topic-hierarchy'
 
@@ -63,42 +61,6 @@ def sync(ctx, logfile, verbosity):
 
     if USERDIR.exists():
         shutil.rmtree(USERDIR)
-
-    LOGGER.debug('Caching WCMP 1.3 artifacts')
-    WCMP1_FILES.mkdir(parents=True, exist_ok=True)
-
-    LOGGER.debug(f'Downloading WCMP1 schemas and codelists to {WCMP1_FILES}')  # noqa
-    ZIPFILE_URL = 'https://wis.wmo.int/2011/schemata/iso19139_2007/19139.zip'  # noqa
-    FH = io.BytesIO(urlopen_(ZIPFILE_URL).read())
-    with zipfile.ZipFile(FH) as z:
-        z.extractall(WCMP1_FILES)
-    CODELIST_URL = 'https://wis.wmo.int/2012/codelists/WMOCodeLists.xml'
-
-    schema_filename = WCMP1_FILES / 'WMOCodeLists.xml'
-
-    with schema_filename.open('wb') as f:
-        f.write(urlopen_(CODELIST_URL).read())
-
-    # because some ISO instances ref both gmd and gmx, create a
-    # stub xsd in order to validate
-    SCHEMA = etree.Element('schema',
-                           elementFormDefault='qualified',
-                           version='1.0.0',
-                           nsmap={
-                               None: 'http://www.w3.org/2001/XMLSchema'
-                           })
-
-    schema_wrapper_filename = WCMP1_FILES / 'iso-all.xsd'
-
-    with schema_wrapper_filename.open('wb') as f:
-        for uri in ['gmd', 'gmx']:
-            namespace = f'http://www.isotc211.org/2005/{uri}'
-            schema_location = f'schema/{uri}/{uri}.xsd'
-
-            etree.SubElement(SCHEMA, 'import',
-                             namespace=namespace,
-                             schemaLocation=schema_location)
-        f.write(etree.tostring(SCHEMA, pretty_print=True))
 
     LOGGER.debug('Caching WCMP2 artifacts')
     LOGGER.debug(f'Downloading WCMP2 schema to {WCMP2_FILES}')
