+++ requirements.txt
@@ -1,6 +1,5 @@
 beautifulsoup4
 click
 jsonschema
-lxml
 pyspellchecker
 python-dateutil
