--- tests/data/not-wcmp.xml
@@ -1,3 +1,0 @@
-<element>
-   <foo>bar</foo>
-</element>
